refactor(tests): consolidate _insert_target into shared _targets_helpers
Moves the duplicated async _insert_target helper from
test_api_v1_targets.py and test_targets_audit_integration.py into the
shared _targets_helpers module, eliminating the structural duplication
that triggered SonarCloud's 3.3% duplication gate on PR #360.
