+++ backend/tests/_targets_helpers.py
@@ -12,6 +12,7 @@ Import into test modules that exercise the /api/v1/targets router:
         _build_app,
         _admin_token,
         _operator_token,
+        _insert_target,
     )
 
 The three autouse fixtures must be imported at module level so pytest
@@ -20,7 +21,9 @@ registers them as autouse for every test in the importing module.
 
 from __future__ import annotations
 
+import uuid
 from collections.abc import Iterator
+from datetime import UTC, datetime
 from typing import Any
 
 import pytest
@@ -31,6 +34,8 @@ from meho_backplane.audit import AuditMiddleware
 from meho_backplane.auth.jwt import clear_jwks_cache
 from meho_backplane.auth.operator import TenantRole
 from meho_backplane.connectors.registry import clear_registry
+from meho_backplane.db.engine import get_sessionmaker
+from meho_backplane.db.models import Target as TargetORM
 from meho_backplane.middleware import RequestContextMiddleware
 from meho_backplane.settings import get_settings
 
@@ -87,3 +92,31 @@ def _operator_token(key: Any, tenant_id: str = DEFAULT_TENANT_ID) -> str:
     return mint_token(
         key, sub="op-1", tenant_role=TenantRole.OPERATOR.value, tenant_id=tenant_id
     )
+
+
+async def _insert_target(**kwargs: Any) -> TargetORM:
+    """Insert a TargetORM row directly via the test sessionmaker."""
+    defaults: dict[str, Any] = {
+        "id": uuid.uuid4(),
+        "tenant_id": uuid.UUID(DEFAULT_TENANT_ID),
+        "name": "default-target",
+        "product": "ssh",
+        "host": "10.0.0.1",
+        "aliases": [],
+        "port": None,
+        "fqdn": None,
+        "secret_ref": None,
+        "auth_model": "shared_service_account",
+        "vpn_required": False,
+        "extras": {},
+        "notes": None,
+        "created_at": datetime.now(UTC),
+        "updated_at": datetime.now(UTC),
+    }
+    defaults.update(kwargs)
+    t = TargetORM(**defaults)
+    sm = get_sessionmaker()
+    async with sm() as session:
+        session.add(t)
+        await session.commit()
+    return t
