+++ backend/tests/test_api_v1_targets.py
@@ -32,8 +32,6 @@ import respx
 from fastapi.testclient import TestClient
 
 from meho_backplane.connectors.schemas import ProbeResult
-from meho_backplane.db.engine import get_sessionmaker
-from meho_backplane.db.models import Target as TargetORM
 
 from ._oidc_jwt_helpers import (
     DEFAULT_TENANT_ID,
@@ -45,6 +43,7 @@ from ._targets_helpers import (
     _admin_token,
     _build_app,
     _empty_connector_registry,  # noqa: F401
+    _insert_target,
     _isolated_jwks_cache,  # noqa: F401
     _operator_token,
     _settings_env,  # noqa: F401
@@ -58,35 +57,6 @@ from ._targets_helpers import (
 @pytest.fixture
 def client() -> Iterator[TestClient]:
     yield TestClient(_build_app())
-
-
-# ---------------------------------------------------------------------------
-# DB helpers — direct inserts without going through the API
-# ---------------------------------------------------------------------------
-
-
-async def _insert_target(**kwargs: Any) -> TargetORM:
-    """Insert a TargetORM row directly via the test sessionmaker."""
-    defaults: dict[str, Any] = {
-        "id": uuid.uuid4(),
-        "tenant_id": uuid.UUID(DEFAULT_TENANT_ID),
-        "name": "default-target",
-        "product": "ssh",
-        "host": "10.0.0.1",
-        "aliases": [],
-        "vpn_required": False,
-        "auth_model": "shared_service_account",
-        "extras": {},
-        "created_at": datetime.now(UTC),
-        "updated_at": datetime.now(UTC),
-    }
-    defaults.update(kwargs)
-    t = TargetORM(**defaults)
-    sm = get_sessionmaker()
-    async with sm() as session:
-        session.add(t)
-        await session.commit()
-    return t
 
 
 # ---------------------------------------------------------------------------
