+++ backend/tests/test_targets_audit_integration.py
@@ -21,9 +21,7 @@ from __future__ import annotations
 
 import uuid
 from collections.abc import AsyncIterator
-from datetime import UTC, datetime
 from pathlib import Path
-from typing import Any
 
 import pytest
 import respx
@@ -43,7 +41,6 @@ from meho_backplane.db.engine import (
 )
 from meho_backplane.db.migrations import alembic_config
 from meho_backplane.db.models import AuditLog
-from meho_backplane.db.models import Target as TargetORM
 from meho_backplane.targets.resolver import (
     TargetNotFoundError,
     resolve_target,
@@ -59,6 +56,7 @@ from ._oidc_jwt_helpers import (
 from ._targets_helpers import (
     _build_app,
     _empty_connector_registry,  # noqa: F401
+    _insert_target,
     _isolated_jwks_cache,  # noqa: F401
     _settings_env,  # noqa: F401
 )
@@ -98,38 +96,6 @@ async def isolated_engine(
     finally:
         await dispose_engine()
         reset_engine_for_testing()
-
-
-# ---------------------------------------------------------------------------
-# DB helpers
-# ---------------------------------------------------------------------------
-
-
-async def _insert_target(**kwargs: Any) -> TargetORM:
-    defaults: dict[str, Any] = {
-        "id": uuid.uuid4(),
-        "tenant_id": _DEFAULT_TENANT_UUID,
-        "name": "test-target",
-        "aliases": [],
-        "product": "rke2",
-        "host": "10.0.0.1",
-        "port": None,
-        "fqdn": None,
-        "secret_ref": None,
-        "auth_model": "shared_service_account",
-        "vpn_required": False,
-        "extras": {},
-        "notes": None,
-        "created_at": datetime.now(UTC),
-        "updated_at": datetime.now(UTC),
-    }
-    defaults.update(kwargs)
-    t = TargetORM(**defaults)
-    sm = get_sessionmaker()
-    async with sm() as session:
-        session.add(t)
-        await session.commit()
-    return t
 
 
 async def _fetch_audit_rows(eng: AsyncEngine) -> list[AuditLog]:
